+++ frontend/src/features/users/stores/reducers/signinReducer.ts
@@ -6,14 +6,14 @@ import {
   pendingReducer,
   fulfilledReducer,
   rejectedReducer,
-} from "@/features/users/stores/reducers/authReducer";
+} from "@/features/users/stores/reducers/authExtraReducer";
 import type {
   AuthState,
   SigninInput,
   AuthResponse,
 } from "@/features/users/types/authTypes";
 
-export const signin = createAsyncThunk<AuthResponse, SigninInput>(
+export const signinAction = createAsyncThunk<AuthResponse, SigninInput>(
   "auth/signin",
   async (input) => {
     return signinApi(input);
@@ -24,7 +24,7 @@ export const buildSigninExtraReducer = (
   builder: ActionReducerMapBuilder<AuthState>
 ) => {
   builder
-    .addCase(signin.pending, pendingReducer)
-    .addCase(signin.fulfilled, fulfilledReducer)
-    .addCase(signin.rejected, rejectedReducer);
+    .addCase(signinAction.pending, pendingReducer)
+    .addCase(signinAction.fulfilled, fulfilledReducer)
+    .addCase(signinAction.rejected, rejectedReducer);
 };
